WIP move app/smoke test to test/integration/
